pyproject.toml
[project]
name = "last-hub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-markdownx>=4.0.9",
    "ephem>=4.1",
    "requests>=2.31"
]
